[#21][refactor]entity 패키지를 base로 이름 변경
