+++ src/main/java/dev/idion/bladitodo/domain/board/Board.java
@@ -1,6 +1,6 @@
 package dev.idion.bladitodo.domain.board;
 
-import dev.idion.bladitodo.domain.entity.BaseEntity;
+import dev.idion.bladitodo.domain.base.BaseEntity;
 import dev.idion.bladitodo.domain.list.List;
 import dev.idion.bladitodo.domain.log.Log;
 import java.util.ArrayList;
