+++ src/main/java/dev/idion/bladitodo/domain/list/List.java
@@ -1,8 +1,8 @@
 package dev.idion.bladitodo.domain.list;
 
+import dev.idion.bladitodo.domain.base.BaseEntity;
 import dev.idion.bladitodo.domain.board.Board;
 import dev.idion.bladitodo.domain.card.Card;
-import dev.idion.bladitodo.domain.entity.BaseEntity;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import javax.persistence.Entity;
